+++ manage.py
@@ -1,50 +1,44 @@
 #!env/bin/python
 
 import os
-import subprocess
-import logging
 
-from flask_script import Manager, Server
-from flask_migrate import Migrate, MigrateCommand
-from whitenoise import WhiteNoise
+from flask_script import Command, Manager, Server
 
 from memegen.settings import get_config
 from memegen.app import create_app
 
 
+class Validate(Command):
+    """Checks for issues in all templates."""
+
+    # pylint: disable=method-hidden
+
+    def run(self):
+        if app.template_service.validate():
+            return 0
+        else:
+            return 1
+
+
+def find_assets():
+    """Yield paths for all static files and templates."""
+    for name in ['static', 'templates']:
+        directory = os.path.join(app.config['PATH'], name)
+        for entry in os.scandir(directory):
+            if entry.is_file():
+                yield entry.path
+
+
 # Select app configuration from the environment
-config = get_config(os.getenv('CONFIG', 'dev'))
+config = get_config(os.getenv('FLASK_CONFIG', 'dev'))
 
 # Build the app using configuration from the environment
-_app = create_app(config)
-
-# Populate unset environment variables
-for name, command in [
-    ('DEPLOY_DATE', "TZ=America/Detroit date '+%F %T'"),
-]:
-    output = subprocess.run(command, shell=True, stdout=subprocess.PIPE,
-                            universal_newlines=True).stdout.strip() or "???"
-    os.environ[name] = os.getenv(name, output)
+app = create_app(config)
 
 # Configure the command-line interface
-manager = Manager(_app)
-
-
-@manager.command
-def validate():
-    if _app.template_service.validate():
-        return 0
-    else:
-        return 1
-
-
-manager.add_command('server', Server(host='0.0.0.0'))
-manager.add_command('db', MigrateCommand)
-
-
-app = WhiteNoise(_app)
-app.add_files("data/images")
-app.add_files("memegen/static")
+manager = Manager(app)
+manager.add_command('validate', Validate())
+manager.add_command('run', Server(host='0.0.0.0', extra_files=find_assets()))
 
 
 if __name__ == '__main__':