+++ memegen/routes/latest.py
@@ -1,19 +1,19 @@
-from flask import Blueprint, render_template
+from flask import Blueprint, render_template, current_app
 from flask_api.decorators import set_renderers
 from flask_api.renderers import HTMLRenderer
 
-from ._common import route, get_tid
+from ._utils import route
 
 
-blueprint = Blueprint('latest', __name__, url_prefix="/latest")
+blueprint = Blueprint('latest-page', __name__)
 
 
-@blueprint.route("")
+@blueprint.route("/latest")
 @set_renderers(HTMLRenderer)
 def get():
     return render_template(
         'latest.html',
-        src=route('image.get_latest'),
-        refresh=10,
-        ga_tid=get_tid(),
+        srcs=[route('image.get_latest', index=i) for i in range(30)],
+        refresh=60,
+        config=current_app.config,
     )